[REFACTOR] S3 로직 external package로 이동
