+++ src/main/java/site/katchup/katchupserver/external/s3/AwsS3Config.java
@@ -1,4 +1,4 @@
-package site.katchup.katchupserver.config;
+package site.katchup.katchupserver.external.s3;
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
