+++ src/main/java/site/katchup/katchupserver/external/s3/PreSignedUrlVO.java
@@ -1,0 +1,12 @@
+package site.katchup.katchupserver.external.s3;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor(staticName = "of")
+public class PreSignedUrlVO {
+    private String fileName;
+    private String fileUploadDate;
+    private String preSignedUrl;
+}
